Sets up jetpack build as a subcommand
