+++ tools/cli/bin/jetpack
@@ -3,9 +3,9 @@
 try {
 	// eslint-disable-next-line no-global-assign
 	require = require( 'esm' )( module /*, options*/ );
-	var args = require( 'args' );
 	const child_process = require( 'child_process' );
 	child_process.execSync( 'jetpack-clicheck202012', { stdio: 'ignore' } );
+	require( '../index' ).cli();
 } catch ( error ) {
 	if ( error.code === "MODULE_NOT_FOUND" ) { // if yarn install hasn't been run, esm require will fail here.
 		console.log( 'Something is missing from your install. Please run `yarn cli-init` and try again.' );
@@ -16,20 +16,4 @@ try {
 		console.log( 'Something unexpected happened. See error above.' );
 	}
 	process.exit( 1 );
-} finally {
-	args.option( 'verbose', 'Display verbose output.' )
-		.command( 'build', 'Builds Jetpack projects.' );
-
-	const flags = args.parse( process.argv );
-	console.log( flags );
 }
-
-/**
-try {
-	require( '../index' ).cli( process.argv );
-} catch ( error ) {
-
-}
- **/
-
-// require = require( 'esm' )( module /*, options*/ );
