+++ tools/cli/builder.js
@@ -1,10 +1,23 @@
+/**
+ * External imports.
+ */
 import child_process from 'child_process';
 import chalk from 'chalk';
+import path from 'path';
+
+/**
+ * Internal imports.
+ */
 import { chalkJetpackGreen } from './helpers/styling';
+import { promptForProject } from './helpers/promptForProject';
+import { cliFunctions } from './helpers/cliFunctions';
+
 // eslint-disable-next-line no-console
 const log = console.log;
 
 /**
+ * Relays build commands to a particular project.
+ *
  * @param options
  */
 export async function builder( options ) {
@@ -12,7 +25,6 @@ export async function builder( options ) {
 		...options,
 		targetDirectory: options.targetDirectory || process.cwd(),
 	};
-
 	switch ( options.project ) {
 		case 'plugins/jetpack':
 			log(
@@ -21,7 +33,8 @@ export async function builder( options ) {
 						'Go ahead and sit back. Relax. This will take a few minutes.'
 				)
 			);
-			child_process.spawnSync( 'yarn', [ 'build-jetpack' ], {
+			child_process.spawnSync( 'yarn', [ 'build' ], {
+				cwd: path.resolve( 'projects/plugins/jetpack' ), // If I can get options.project to work...
 				shell: true,
 				stdio: 'inherit',
 			} );
@@ -29,6 +42,21 @@ export async function builder( options ) {
 		default:
 			log( chalk.yellow( 'This project does not have a build step defined.' ) );
 	}
-
 	return true;
+}
+
+/**
+ * Entry point for the CLI.
+ */
+export async function cli() {
+	const cli = cliFunctions();
+	// Add cli. commands here to string together options.
+
+	let options = cli.parse( process.argv, { version: false } );
+
+	options = await promptForProject( options );
+	if ( options.verbose ) {
+		console.log( options );
+	}
+	await builder( options );
 }
